Simplify cube movment/rotation/scaling system
Instead of maintaining offsets on the cubes themselves, move offsets
to mouse movement handler function, that way system is entirely
contained within that function. Now cube move/rotate/scale functions
take a simple change vector and apply it, instead of a complete mouse
movement vector that needs to be offset.